backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
djangorestframework==3.12.4
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
pytz==2021.1
requests==2.26.0
simplejson==3.17.4
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
